Add .bin extension to avoid naming collision between binary and folder
On Linux and Mac the resulting binary would be named "magnet" but it
can't create that file because a directory called "magnet" already
exists in the same place.  In Windows this problem doesn't happen
because the output is magnet.exe.  Adding the .bin extension for Linux
and Mac avoid the naming collision with the existing directory with no
negative effects because the file extension doesn't play an important
role and .bin works fine for this case.